pyproject.toml
[tool.poetry]
name = "kafka-first-steps"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "kafka_first_steps"}]

[tool.poetry.dependencies]
python = "^3.8"
kafka-python = "^2.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
